[doc] add jsdoc documentation
