+++ src/index.ts
@@ -5,6 +5,8 @@ import type {
   DeclinedType,
   PrincipalPartsType,
 } from './types.ts'
+import type { InfinitiveType } from '~conjugators/InfinitiveConjugator.ts'
+import type { ParticipleType } from '~conjugators/ParticipleDecliner.ts'
 import type { PadalyvisType } from '~conjugators/PadalyvisInflector.ts'
 import type { PusdalyvisType } from '~conjugators/PusdalyvisDecliner.ts'
 import type { BudinysType } from '~conjugators/BudinysInflector.ts'
@@ -14,7 +16,9 @@ export type {
   BudinysType,
   ConjugationType,
   DeclinedType,
+  InfinitiveType,
   PadalyvisType,
+  ParticipleType,
   PrincipalPartsType,
   PusdalyvisType,
 }
