+++ test/Verb.test.ts
@@ -7,8 +7,73 @@ import type { InflectorInterface } from '~conjugators/Inflector.ts'
 import type ParticipleDecliner from '~conjugators/ParticipleDecliner.ts'
 import type { ParticipleType } from '~conjugators/ParticipleDecliner.ts'
 import type { BudinysType } from '~conjugators/BudinysInflector.ts'
+import {
+  threeRootsError,
+  unmatchingPrefixesError,
+  unmatchingReflexivesError,
+} from '~src/errors.ts'
+import type { PrincipalPartsType } from '~src/types.ts'
 
 describe('Verb', () => {
+  describe('constructor', () => {
+    const principalParts: PrincipalPartsType = ['rinkti', 'renka', 'rinko']
+    it('should get correct basic input', () => {
+      const verb = new Verb('rinkti-renka-rinko')
+      expect(verb.principalParts).toStrictEqual(principalParts)
+      expect(verb.isReflexive).toStrictEqual(false)
+      expect(verb.prefix).toStrictEqual(undefined)
+      assertEqualVerbs(new Verb(principalParts), verb)
+    })
+    it('should get correct reflexive input', () => {
+      const verb = new Verb('rinktis-renkas-rinkosi')
+      expect(verb.principalParts).toStrictEqual(principalParts)
+      expect(verb.isReflexive).toStrictEqual(true)
+      expect(verb.prefix).toStrictEqual(undefined)
+      assertEqualVerbs(new Verb(principalParts, { reflexive: true }), verb)
+    })
+    it('should get correct prefixed input', () => {
+      const verb = new Verb('ne=rinkti-ne=renka-ne=rinko')
+      expect(verb.principalParts).toStrictEqual(principalParts)
+      expect(verb.isReflexive).toStrictEqual(false)
+      expect(verb.prefix).toStrictEqual('ne')
+      assertEqualVerbs(new Verb(principalParts, { prefix: 'ne' }), verb)
+    })
+    it('should get correct prefixed and reflexed input', () => {
+      const verb = new Verb('ne=rinktis-ne=renkasi-ne=rinkos')
+      expect(verb.principalParts).toStrictEqual(principalParts)
+      expect(verb.isReflexive).toStrictEqual(true)
+      expect(verb.prefix).toStrictEqual('ne')
+      assertEqualVerbs(
+        new Verb(principalParts, { prefix: 'ne', reflexive: true }),
+        verb,
+      )
+    })
+    it('should get correct prefix with reflexive input', () => {
+      const verb = new Verb('nesi=rinkti-nesi=renka-nesi=rinko')
+      expect(verb.principalParts).toStrictEqual(principalParts)
+      expect(verb.isReflexive).toStrictEqual(true)
+      expect(verb.prefix).toStrictEqual('ne')
+      assertEqualVerbs(
+        new Verb(principalParts, { prefix: 'ne', reflexive: true }),
+        verb,
+      )
+    })
+    it('should throw when prefixes mismatch', () => {
+      expect(() => new Verb('rinkti-ne=renka-ne=rinko')).toThrow(
+        unmatchingPrefixesError,
+      )
+    })
+    it('should throw when reflexive mismatches', () => {
+      expect(() => new Verb('rinktis-renka-rinko')).toThrow(
+        unmatchingReflexivesError,
+      )
+    })
+    it('it should throw when principal part count is not 3', () => {
+      expect(() => new Verb('')).toThrow(threeRootsError)
+      expect(() => new Verb('dėti-dedu')).toThrow(threeRootsError)
+      expect(() => new Verb('dėti-dedu-dėjo-dėdavo')).toThrow(threeRootsError)
+    })
+  })
   describe('conjugations', () => {
     assertTense(
       Verb.pastFrequentativeIndicative,
@@ -298,3 +363,9 @@ describe('Verb', () => {
     }
   })
 })
+
+function assertEqualVerbs(a: Verb, b: Verb) {
+  expect(a.principalParts).toStrictEqual(b.principalParts)
+  expect(a.prefix).toStrictEqual(b.prefix)
+  expect(a.isReflexive).toStrictEqual(b.isReflexive)
+}
